requirements.txt
pdfminer.six>=20221105
python-dotenv>=1.0.0
tqdm>=4.65.0
openai>=1.0.0
